Update metrics_fast.py
Hey, thanks for the toolbox, it's a major contribution for my research! I found some deviations in calculating TM-30 metrics while using the metrics_fast methods in the mixed reference area from 4000 to 5000 K. The problem seems to be that the index for mixed_range  in line 54 should be the upper bound [1] but is set to the lower one [0]. So as a result, it is either BB or DL but never the mix of both.
Best regards!